Added X,Y pixel coordinate label for areas of interest thumbnail, Added "Max Object Area (px)" field to main window
Added X,Y pixel coordinate label for an area of interest thumbnail in the viewer. The X,Y pixel coordinate is useful when using ADIAT to examine images from CAP (Civil Air Patrol) flights or other high resolution images. The pixel coordinate allows the user to find the location in the image if they are viewing the image in some other program.
Added "Max Object Area (px)" field to the main window. The maximum object area field is useful to prevent false positives. For example if we know our subject is a human sized person we can set the max object area to the number of pixels that would represent the maximum size of a human. That way any areas of interest that are identified that are larger then that size would not be included in the areas of interest list.