Copy XMP data to output file.
This code adds two functions to MetaDataHelper.py
get_xmp_data: This function checks the file passed to see if it contains xmp data by looking for the "<?xpacket begin" header. If it is found it returns the data else it returns nothing.
set_xmp_data: This function will copy the xmp data passed to it into the file that is passed to it.

In Algorithm.py when copying meta data to an output file the code will also check to see if xmp data exists and it does it copies it to the output file.